+++ ldso/ldso/i386/ld_sysdep.h
@@ -25,8 +25,8 @@
  */
 #define INIT_GOT(GOT_BASE,MODULE) \
 {				\
-  GOT_BASE[2] = (int) _dl_linux_resolve; \
-  GOT_BASE[1] = (int) MODULE; \
+  GOT_BASE[2] = (unsigned long) _dl_linux_resolve; \
+  GOT_BASE[1] = (unsigned long) MODULE; \
 }
 
 /*
